% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_bill.R
\docType{class}
\name{bill-class}
\alias{bill-class}
\title{定义一个单据的基本类型}
\value{
没有返回值
}
\description{
定义一个单据的基本类型
}
\section{Slots}{

\describe{
\item{\code{FInterId}}{integer. 单据内码}

\item{\code{FNumber}}{character. 单据代码}

\item{\code{FName}}{character.  单据名称}
}}

\examples{
 不需要示例
}
