% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDate.R
\docType{methods}
\name{FDate,salesOrderHead-method}
\alias{FDate,salesOrderHead-method}
\title{取用于对象于提供相应的数据}
\usage{
\S4method{FDate}{salesOrderHead}(X)
}
\arguments{
\item{bill}{基类bill}
}
\value{
返回相应的数据
}
\description{
取用于对象于提供相应的数据
}
\examples{
FDate(data);
}
