% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FInterId.R
\docType{methods}
\name{FInterId,bill-method}
\alias{FInterId,bill-method}
\title{取用于对象于提供相应的数据}
\usage{
\S4method{FInterId}{bill}(X)
}
\arguments{
\item{bill}{基类bill}
}
\value{
返回相应的数据
}
\description{
取用于对象于提供相应的数据
}
\examples{
FInterId(data);
}
