% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FName.R
\docType{methods}
\name{FName,bill-method}
\alias{FName,bill-method}
\title{取用于对象于提供相应的数据}
\usage{
\S4method{FName}{bill}(X)
}
\arguments{
\item{bill}{基类bill}
}
\value{
返回相应的数据
}
\description{
取用于对象于提供相应的数据
}
\examples{
FName(data);
}
