% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FNumber.R
\docType{methods}
\name{FNumber,bill-method}
\alias{FNumber,bill-method}
\title{取用于对象于提供相应的数据}
\usage{
\S4method{FNumber}{bill}(X)
}
\arguments{
\item{bill}{基类bill}
}
\value{
返回相应的数据
}
\description{
取用于对象于提供相应的数据
}
\examples{
FNumber(data);
}
