% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FNumber.R
\name{FNumber<-}
\alias{FNumber<-}
\title{设置FNumber替代的泛型函数}
\usage{
FNumber(X) <- value
}
\arguments{
\item{X}{对象}

\item{value}{替代值}
}
\value{
返回结果
}
\description{
设置FNumber替代的泛型函数
}
\examples{
FNumber(data) <-value;
}
