% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAcctData_L1.R
\name{getAcctData_L1}
\alias{getAcctData_L1}
\title{从财务提供的科目初始数据中提供1级发生数}
\usage{
getAcctData_L1(netstore_3year_balance)
}
\arguments{
\item{netstore_3year_balance}{数据框报表}
}
\value{
返回处理后的数据框
}
\description{
从财务提供的科目初始数据中提供1级发生数
}
\examples{
getAcctData_L1(DATA);
}
