% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPLData.R
\name{getPLData}
\alias{getPLData}
\title{从所有一级科目的发生数据中提取出损益的发生数
处理口径的数据，进行了过滤处理}
\usage{
getPLData(data)
}
\arguments{
\item{data}{来源于getAcctData_L1 function处理后的数据}
}
\value{
返回一个数据框
}
\description{
从所有一级科目的发生数据中提取出损益的发生数
处理口径的数据，进行了过滤处理
}
\examples{
 getPLData(data);
}
