% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huanBiCalc.R
\name{huanBiCalc}
\alias{huanBiCalc}
\title{根据一个时间序列数据计算环比上一期金额，差异，增长百分比}
\usage{
huanBiCalc(value)
}
\arguments{
\item{data}{一个数据值的数组即可，不需要是数据框。}
}
\value{
返回一个数据框；
}
\description{
根据一个时间序列数据计算环比上一期金额，差异，增长百分比
}
\examples{
 huanBiCalc(x);
}
