% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_salesOrderHead.R
\docType{class}
\name{salesOrderHead-class}
\alias{salesOrderHead-class}
\title{定义销售订单的表头字段}
\value{
没有返回值
}
\description{
定义销售订单的表头字段
}
\section{Slots}{

\describe{
\item{\code{FDate}}{character.单据日期}
}}

\examples{
不做示例
}
