% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salesOrderHead.R
\name{salesOrderHead}
\alias{salesOrderHead}
\title{实例化salesOrderHead一个娄为对象}
\usage{
salesOrderHead(data)
}
\arguments{
\item{data}{数据框,字名名必须为FInterId,FNumber,FName,FDate}
}
\value{
返回实例化对象
}
\description{
实例化salesOrderHead一个娄为对象
}
\examples{
salesOrderHead(data.frame(FInterId,FNumber,FName,FDate));
}
